requirements.txt
requests>=2.31.0
python-gitlab>=4.0.0
gitpython>=3.1.0
click>=8.0.0
